+++ internal/metadata_repository/reporter_client_factory_impl.go
@@ -3,17 +3,22 @@ package metadata_repository
 import (
 	"context"
 
+	"google.golang.org/grpc"
+
 	"github.com/kakao/varlog/internal/storagenode/reportcommitter"
 	"github.com/kakao/varlog/proto/varlogpb"
 )
 
 type reporterClientFactory struct {
+	grpcDialOptions []grpc.DialOption
 }
 
-func NewReporterClientFactory() *reporterClientFactory {
-	return &reporterClientFactory{}
+func NewReporterClientFactory(grpcDialOptions ...grpc.DialOption) *reporterClientFactory {
+	return &reporterClientFactory{
+		grpcDialOptions: grpcDialOptions,
+	}
 }
 
 func (rcf *reporterClientFactory) GetReporterClient(ctx context.Context, sn *varlogpb.StorageNodeDescriptor) (reportcommitter.Client, error) {
-	return reportcommitter.NewClient(ctx, sn.Address)
+	return reportcommitter.NewClient(ctx, sn.Address, rcf.grpcDialOptions...)
 }